move interoperable data formats used in serialization into its own common directory
