+++ client/src/scenes/ReplayBrowser.lua
@@ -2,7 +2,7 @@ local Scene = require("client.src.scenes.Scene")
 local input = require("common.lib.inputManager")
 local GraphicsUtil = require("client.src.graphics.graphics_util")
 local fileUtils = require("client.src.FileUtils")
-local Replay = require("common.engine.Replay")
+local Replay = require("common.data.Replay")
 local class = require("common.lib.class")
 local GameModes = require("common.engine.GameModes")
 local ReplayGame = require("client.src.scenes.ReplayGame")
