+++ client/src/scores.lua
@@ -1,4 +1,4 @@
-local levelPresets = require("common.engine.LevelPresets")
+local levelPresets = require("common.data.LevelPresets")
 local fileUtils = require("client.src.FileUtils")
 local class = require("common.lib.class")
