+++ client/tests/AttackFileGenerator.lua
@@ -1,7 +1,7 @@
 local logger = require("common.lib.logger")
 local separator = package.config:sub(1, 1) --determines os directory separator (i.e. "/" or "\")
 local fileUtils = require("client.src.FileUtils")
-local Replay = require("common.engine.Replay")
+local Replay = require("common.data.Replay")
 
 -- Helper debug functions for analyzing a set of replays and outputting attack files from them.
