+++ common/engine/Match.lua
@@ -12,8 +12,8 @@ local class = require("common.lib.class")
 local logger = require("common.lib.logger")
 local tableUtils = require("common.lib.tableUtils")
 local GameModes = require("common.engine.GameModes")
-local Replay = require("common.engine.Replay")
-local ReplayPlayer = require("common.engine.ReplayPlayer")
+local Replay = require("common.data.Replay")
+local ReplayPlayer = require("common.data.ReplayPlayer")
 local Signal = require("common.lib.signal")
 local SimulatedStack = require("common.engine.SimulatedStack")
 local Stack = require("common.engine.Stack")
