+++ common/engine/checkMatches.lua
@@ -2,7 +2,7 @@ local logger = require("common.lib.logger")
 local tableUtils = require("common.lib.tableUtils")
 local PanelGenerator = require("common.engine.PanelGenerator")
 local consts = require("common.engine.consts")
-local LevelData = require("common.engine.LevelData")
+local LevelData = require("common.data.LevelData")
 local prof = require("common.lib.jprof.jprof")
 require("table.clear")
